% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{distance}
\alias{distance}
\title{distance}
\usage{
distance(x, y, data = NULL, method = c("euclidean", "manhattan", "canberra",
  "minkowski"), p = 2L)
}
\arguments{
\item{x}{The x vector.}

\item{y}{The y vector.}

\item{data}{If data is provided x and y will be taken from data.}

\item{method}{One of euclidean, manhattan, canberra, minkowski.}

\item{p}{Only used if method is minkowski.}
}
\value{
The distance between x and y.
}
\description{
Calculates the distance between two vectors
}
\details{
Calculates the distance between two vectors. Euclidean, manhattan, canberra and minkowski are
supported.
}
\examples{
set.seed(10)
distance(runif(100), runif(100))

}

