% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_scale.R
\name{feature_scale}
\alias{feature_scale}
\title{Scales vector between 0 and 1}
\usage{
feature_scale(x)
}
\arguments{
\item{x}{The atomic vector.}
}
\value{
The scaled vector.
}
\description{
Scales vector between 0 and 1
}
\details{
Feature scaling is a technique used to transform a set of values between 0 and 1.
}
\examples{
feature_scale(0:10)

}

