% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_cor.R
\name{remove_cor}
\alias{remove_cor}
\title{Removes Correlated Variables}
\usage{
remove_cor(data, threshold = 0.8)
}
\arguments{
\item{data}{The data.frame to check for multicollinearity.}

\item{threshold}{Correlation threshold.}
}
\value{
The name of the correlated variables.
}
\description{
Removes Correlated Variables
}
\details{
Calculates the correlation matrix using \code{cor}. Then for each variable checks the correlation
against all the other variables and removes the ones that are highly correlated according to the
set threshold. Repeats the process until there are no correlated variables left.
}
\examples{
remove_cor(mtcars)

}

