% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_ifelse.R
\name{safe_ifelse}
\alias{safe_ifelse}
\title{A safe version of ifelse that preserves class}
\usage{
safe_ifelse(test, yes, no)
}
\arguments{
\item{test}{An object which can be coerced to logical mode.}

\item{yes}{return values for true elements of test.}

\item{no}{return values for false elements of test.}
}
\value{
Same as ifelse but preserving the class (works with dates).
}
\description{
A safe version of ifelse that preserves class
}
\details{
A safe version of ifelse that preserves class. Taken from 
\url{http://stackoverflow.com/questions/6668963/how-to-prevent-ifelse-from-turning-date-objects-into-numeric-objects}
}
\examples{
dates <- as.Date(c('2011-01-01', '2011-01-02', '2011-01-03', '2011-01-04', '2011-01-05'))
safe_ifelse(dates == '2011-01-01', dates - 1, dates)

}

