% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switchv.R
\name{switchv}
\alias{switchv}
\title{A vectorised version of switch}
\usage{
switchv(EXPR, type = character(1), ...)
}
\arguments{
\item{EXPR}{An expression evaluating to a number or a character string.}

\item{type}{\code{switchv} uses \code{vapply} so this is the type returned from each iteration.
Same as FUN.VALUE argument in \code{vapply}.}

\item{...}{the list of alternatives.}
}
\value{
Same as \code{switch} but used on a vector.
}
\description{
A vectorised version of switch
}
\examples{
myvec <- c('apple', 'cabbage')
switchv(myvec, 
        apple = 'fruit',
        cabbage = 'veggetable')

}

